Split app.py and added mail routes
